refactor: create more seperate structure for two types of import
