--- backend/data_import/excel/core/config.py
@@ -1,14 +1,0 @@
-from enum import Enum
-from typing import ClassVar, final
-
-
-@final
-class ExcelDirectory(Enum):
-    E8 = "E8_data"
-    EM = "EM_data"
-
-
-@final
-class ExcelFile:
-    SHEET_NAME = "SAS"
-    HEADER: ClassVar[list[int]] = [0, 1]
